Generated infrastructure model
### DESCRIPTION

This pull request was automatically generated from the terraform code within this repo, and adds a UML-deployment diagram to your documentation. Let me know what you think in the comments.

### Feedback

This diagram was created by a tool developed by the Faculty of Engineering of the University of Porto (Portugal), in the context of a research project. 
The project has no commercial goal. Whether you decide on accepting or rejecting this pull request, we would like to invite you to answer a 5min survey, which will help plan future research and improvements to the tool.

Please answer the survey, [here](https://forms.gle/DZSSM1JkBaJefnDGA).
Thank you.